team_generator/pyproject.toml
[tool.poetry]
name = "team-generator"
version = "0.1.0"
description = ""
authors = ["Renan Santana <renan_ota@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.6"
jinja2 = "^3.1.2"
sqlalchemy = "^2.0.17"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
